+++ RELEASE_NOTES.md
@@ -10,6 +10,7 @@
 - #135 Fixes put merge error on empty folder when it exists in destination
 - Fixes error message when undoing copy of empty folder
 - Fixes Find input error not going away after using Go To Next Find Match
+- Fixes Back and Forward to not leave inaccessible paths on stack
 - Fixes drive not being selected when opening parent
 - Fixes Refresh window on focus with recursive search to refresh current directory
 - Undo/redo now reports status for each iteration so they can be seen in history